Add index to setup and refactor out
